backend/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy>=2.0
pymysql
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
pydantic[email]
